Move components out of Pages
