+++ components/Wordiables.tsx
